drm: vc4: dsi: Stop the pixel valve (crtc) before the DSI block
The DSI block wants the PV to be disabled before the DSI block
in order to ensure that all FIFOs are emptied correctly.

With the DSI driver being in the bridge chain, it is fully disabled
before the crtc is disabled afterwards, so add a call to the crtc
atomic_disable from the DSI atomic_disable.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>