media: i2c: Correct definition for OV7251_TIMING_MIN_VTS
Currently there are defines for OV7251_TIMING_MIN_VTS and
OV7251_TIMING_MAX_VTS nominally relating to the VTS register
in the sensor.
This register takes a range of the (image height +
offset) to 0xffff, so the definition of OV7251_TIMING_MIN_VTS
being 1 is incorrect.

Rename OV7251_TIMING_MIN_VTS to OV7251_TIMING_VBLANK_MIN as
it is used minimum value for V4L2_CID_VBLANK, which is the
offset part of the minimum value that the VTS register is
programmed with.
The max value is already computed as OV7251_TIMING_MAX_VTS -
image height.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>