dmaengine: bcm2835: Add support for the first DMA interrupt not being 0
Variants of the Broadcom chips about to be introduced have DMA
controllers in the same block with varying capabilities over
the address space they can work with, so they will need to be
split over several dma-controller nodes.
However they are referred to by consecutive numbers, and it would
be useful to maintain that naming, and the coherent masking.

The base address will match the first channel for each grouping
of capability, and the interrupts defined will be defined for
all supported channels on that controller.

Use the first interrupt found to determine the offset in channel
number for the block.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>