media: i2c: Add a driver for the onsemi AR0144 camera sensor
The AR0144 is a 1/4" 1MP CMOS camera sensor from onsemi. The driver
supports both the monochrome and color versions, and both the parallel
and MIPI CSI-2 interfaces. Due to limitations of the test platform, only
the CSI-2 output has been tested.

The driver supports 8-, 10- and 12-bit output formats, analog crop and
skipping. It exposes controls that cover the usual use cases for camera
sensors.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>