drm/vc4: tests: Patch up the vc4 kunit tests
Commit [1] removed the dynamic allocation of drm_modeset_acquire_ctx
structures, with clients expected to allocate them by other means,
such as on the stack. This usage model works less well when tests
are split into separate initialisation and run phases, because
a ctx allocated on the stack in the init function will go out of
scope and likely be overwritten by the run function(s).

As kunit tests seem to be single threaded, make ctx static to avoid
the issue.

[1] commit 30188df0c387 ("drm/tests: Drop drm_kunit_helper_acquire_ctx_alloc()")

Signed-off-by: Phil Elwell <phil@raspberrypi.com>