media: i2c: imx290: Set INCKSEL registers based on link frequency
Whilst the INCKSEL registers are specified against modes in the
datasheet, that is because the 720p mode can be run at a lower
link frequency to the 1080p. The window cropping mode that could
also produce a cropped 720p mode uses the same INCKSEL register
settings as the 1080p mode.

Set the registers based on the link frequency choice rather than
the mode.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>

---
Currently this will have exactly no effect as the mode sets the link
frequency, but that's about to change.