media: i2c: imx290: Allow 720p mode to run at either link frequency
The 720p mode on this sensor is documented as running with a lower
link frequency to the 1080p mode. However there is no requirement
for this as the window cropping mode that could achieve exactly
the same readout area is documented as using the same link frequency
as 1080p.

Make the lower link frequency optional in device tree, and only
switch to it if defined.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>